% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRake.R
\name{adjustSex}
\alias{adjustSex}
\title{Adjust population counts by Sex, when additional prorating is needed before raking.}
\usage{
adjustSex(InputData, CtrlPopTotals, years = NULL, VarSex = "Sex")
}
\arguments{
\item{InputData}{Data variable containing the database to be adjusted. Expects data to be in
data.frame with columns: Year, Type, TypeID, Sex, ages 0:100, TOTAL. If InputData's Age is wide,
it will pivot the age columns long to column "Age". If any BC/overall counts are included (i.e.,
TypeID == 0), they are dropped.}

\item{CtrlPopTotals}{Data variable containing the overall BC database to be adjusted against.
Expects data to be in data.frame with columns: Year, Sex, ages 0:100, TOTAL. The Sex variable is
assumed to be numeric, with the largest Sex value representing Total sex.}

\item{years}{Vector of Years to adjust. Default = NULL. If NULL, the function will iterate over
all years found in InputData (assuming such years are also in CtrlPopTotals).}

\item{VarSex}{Name of Sex variable in both files (e.g., "Sex"). \strong{Note: Sex must be a
numeric variable (e.g., 1,2,3) where the Total is the maximum number (e.g., 3.}}
}
\value{
Database with Sex adjusted to CtrlPopTotals, by TypeID.
\code{\link{dbConvert}}
}
\description{
The APL Estimates Breakdown System (EBS) was unable to rake population estimates of CHSAs
by Age and Sex, unless Sex was adjusted (aka, prorated) before it was raked (even more so than
the prorating included within the raking function). (EBS first used CHSAs in 2019.)
}
\details{
This function takes InputData and adjusts Sex by TypeID counts, compared to CtrlPopTotals (i.e.,
BC age sex population counts). InputData and CtrlPopTotals are assumed to already be in the
environment.
Historically, EBS had been run on the 89 LHAs (local health areas). EBS was first run on CHSAs
in 2019 to break down the 218 CHSA estimate totals to age sex estimates. The results were erratic
in that many young age children (0-5) were removed during the raking process rendering some CHSAs
with very low or no children aged 0-5. This problem was fixed by first prorating the EBS output
by sex (i.e., this adjustSex() function) using the BC level sex totals, then raking as usual.
}
\examples{
\dontrun{  adjustSex(InputData = pop, CtrlPopTotals = BC_AGE_SEX, years = NULL, VarSex = "Sex")  }
\dontrun{  adjustSex(InputData = pop, CtrlPopTotals = BC_AGE_SEX)  }
}
\seealso{
Overall package documentation: \code{\link{dbutils}}()

Other raking helpers: 
\code{\link{add.random.fraction.to.cols}()},
\code{\link{allowNegsnoMargin}()},
\code{\link{calc.cols}()},
\code{\link{multiRake}()},
\code{\link{noNegsneedMargin}()},
\code{\link{noNegsnoMargin}()},
\code{\link{prep.prorate.col}()},
\code{\link{prorate.col}()},
\code{\link{prorate.row}()},
\code{\link{prorate}()},
\code{\link{read.inputs}()},
\code{\link{real.to.int}()},
\code{\link{rename.age.grps}()},
\code{\link{rounded}()}
}
\concept{raking helpers}
