% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRake.R
\name{allowNegsnoMargin}
\alias{allowNegsnoMargin}
\title{Raking Algorithm Function A: when negative values are allowed}
\usage{
allowNegsnoMargin(CurrRow, CurrRow_value, data, n_colGrps, n_rowGrps, RowAdj)
}
\arguments{
\item{CurrRow}{a subset of data with 3 rows and columns: XXX, n_colGrps, Sum, Ctrl_TOTAL,
Diff, adj_value, where XXX is Sex in Part 1, Region in Part 2, and Age in Part 3. The rows
are the one XXX row currently being worked on, Adjustments, and AdjCurrRow.}

\item{CurrRow_value}{a counter to iterate through rows in data}

\item{data}{a dataframe of rows to be iteratively raked with columns: XXX,  n_colGrps, Sum,
Ctrl_TOTAL, Diff, adj_value, where XXX is the variable being raked (Sex in Part 1, Region in Part 2,
Age in Part 3). Rows are the rows of XXX as well as three summary rows: Sum, Ctrl_TOTAL, Diff.}

\item{n_colGrps}{the number of column groups to rake over (e.g., 89 LHAs, number of Age Groups, etc.)}

\item{n_rowGrps}{the number of rows groups to rake over (e.g., 2 Sexes, 89 LHAs, etc.)}

\item{RowAdj}{vector of zeros, of length n_colGrps (e.g., 89 zeros, etc.)}
}
\value{
original dataframe, but with CurrRow's Diff now 0, and any adjustments made to CurrRow
reversed in the row below that has the largest minimum value
}
\description{
When raking, the appropriate algorithm is chosen (negative values allowed or not, margin needed
or not), "whole people" adjustments to be made to selected cells in a row are calculated, then
"residual people" adjustments to be made to selected cells in a row are calculated, then column
control totals are reconciled. Raking is run iteratively row-by-row. Regardless of the algorithm
needed, five arguments are needed. A sixth argument, "needMargin" is only needed for algorithm C
(\code{\link{noNegsneedMargin}}). This is a helper function used in \code{\link{dbRake}} when
negative values ARE allowed.
}
\seealso{
The overall raking function: \code{\link{dbRake}}()

Other raking helpers: 
\code{\link{add.random.fraction.to.cols}()},
\code{\link{calc.cols}()},
\code{\link{noNegsneedMargin}()},
\code{\link{noNegsnoMargin}()},
\code{\link{prep.prorate.col}()},
\code{\link{prorate.col}()},
\code{\link{prorate.row}()},
\code{\link{read.inputs}()},
\code{\link{real.to.int}()},
\code{\link{rounded}()}
}
\concept{raking helpers}
