% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRake.R
\name{calc.cols}
\alias{calc.cols}
\title{Calculate necessary columns}
\usage{
calc.cols(data, temp, VarRow, n_colGrps)
}
\arguments{
\item{data}{a dataframe of inital population counts that need to be adjusted to control totals
(e.g., columns: a region variable ("VarRow"), 1 (for Males), 2 (for Females))}

\item{temp}{a dataframe of control totals to adjust data to (e.g., region control totals)}

\item{VarRow}{the name of the variable to join temp to data (e.g., VarRow for "LHA")}

\item{n_colGrps}{the number of groups to adjust over (e.g., number of Sexes, 2); this is
essentially the number of columns in data minus 1 (for the VarRow column)}
}
\value{
the original dataframe with four new columns (Sum, Ctrl_TOTAL, Diff, adj_value)
}
\description{
calculates necessary columns (Sum, Ctrl_TOTAL, Diff, adj_value) in preparation for prorating rows
or raking. First, it calculates the actual sum (of columns), then adds in VarRow control totals,
then calculates the difference, and finally calculates the adjustment value (difference divided by
number of groups). This is a helper function used in dbRake, and before calling \code{\link{prorate.row}}.
}
\details{
This function is first called in dbRake Part 1 to update initial initial estimates of male/female
regional total values (before prorating rows, and again before raking), then again in Part 2 to
update initial 5 year age group and maximum age group estimates by Sex (before prorating rows,
and again before raking), and finally in Part 3 to update initial single year of age estimates
by Sex (before prorating rows, and again before raking).
}
\seealso{
The overall raking function: \code{\link{dbRake}}()

Other raking helpers: 
\code{\link{add.random.fraction.to.cols}()},
\code{\link{allowNegsnoMargin}()},
\code{\link{noNegsneedMargin}()},
\code{\link{noNegsnoMargin}()},
\code{\link{prep.prorate.col}()},
\code{\link{prorate.col}()},
\code{\link{prorate.row}()},
\code{\link{read.inputs}()},
\code{\link{real.to.int}()},
\code{\link{rounded}()}
}
\concept{raking helpers}
