% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_function_call.R
\name{capture_function_call}
\alias{capture_function_call}
\title{Captures function call}
\usage{
capture_function_call(...)
}
\value{
the function call
}
\description{
A function call and its arguments, including default arguments, are captured and returned as class call
}
