% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_database.R
\name{connect_to_database}
\alias{connect_to_database}
\title{Connect to an internal NEFSC database.}
\usage{
connect_to_database(server, uid)
}
\arguments{
\item{server}{name of the server}

\item{uid}{username of person with permissions}
}
\value{
Object inherited from \code{\link{DBIConnection}}. This object is used to connect
to communicate with the database engine.
}
\description{
A utility function enabling the user to connect to an internal database provided the user has permissions and
posesses a valid username and password. A popup window is used for secure password enntry
}
\details{
For this to work, you need an oracle client installed.
Tested with Oracle instantClient_12_2 installed
Note: if you use 64 bit Rstudio then you need a 64 bit client
Note: if you use 32 bit Rstudio then you need a 32 bit client
}
\section{Warning}{

3 failed attempts and you will be locked out of the system.
}

\examples{
\dontrun{
con <- connectToDatabase(server="name_of_server",uid="individuals_username")
}
}
\seealso{
\code{\link{DBI}},\code{\link{odbc}}
}
