% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sql.R
\name{create_sql}
\alias{create_sql}
\title{called by many functions from several db packages to convert passed arguments to sql statements}
\usage{
create_sql(dataName, fieldName, fieldName2, dataType, defaultSqlStatement)
}
\arguments{
\item{dataName}{Character or numeric. value of argument passed}

\item{fieldName}{Character. Database fieldname name which has value = \code{dataName}}

\item{fieldName2}{Character. Alternative field name for string values of \code{dataName}}

\item{dataType}{Character. Formatting for the \code{dataName} in the database}

\item{defaultSqlStatement}{String. Select statement. (eg. "Select * from db")}
}
\value{
String. SQL statement
}
\description{
utility function to create a SQL statement based on passed arguments
}
\examples{
\dontrun{
create_sql(dataName = "all",fieldName="svspp", fieldname2= "comname", dataType = "\\\%03d",defaultSqlStatement="select * from svdbs.svspecies_list")
}
}
