% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{dbInfo}
\alias{dbInfo}
\title{Population database dimension information}
\usage{
dbInfo(db_path)
}
\arguments{
\item{db_path}{Either vector of database type ("estimates", "projections", "births", "deaths"),
region code and two-character year OR full path to datbase csv file.}
}
\value{
List structure with unique values for Year, Type, TypeID, Region, Age and Sex.
}
\description{
Reads a population database file (population, births, deaths) and returns list with unique
values for each of Year, Type, TypeID, Age and Sex. Database path can be either a full path
to a csv database file with the correct columns, or a vector including the type, region and
year of database to be used through \code{\link{getDBPath}}. For example,
db_path = c("estimates", "HA", "18") would return info for the July 1st 2018 population
estimates database for Local Health Areas.
}
\seealso{
Other database access helpers: 
\code{\link{dbCheck}()},
\code{\link{dbRead}()},
\code{\link{dbWrite}()},
\code{\link{getDBPath}()}
}
\author{
Sebastien Lavoie, (formerly, BC Stats)
}
\concept{database access helpers}
