% Generated by roxygen2 (4.1.1): do not edit by hand
\name{dbIsLive}
\alias{dbIsLive}
\title{Check Connection Validity}
\usage{
dbIsLive(con)
}
\arguments{
  \item{con}{connection object, as returned by \code{\link{dbConnect}}.}
}
\description{
  \code{dbIsLive} tests if a connection is valid, i.e. open.
}

