% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRake.R
\name{dbRake}
\alias{dbRake}
\title{Rake population database}
\usage{
dbRake(
  InputData,
  Ctrl_Pop_Totals,
  Ctrl_Region_Totals = NULL,
  Ctrl_AgeGrps_Totals = NULL,
  VarRegion,
  VarSex,
  VarSexTotal,
  AgeGrpMax = NULL,
  allowNegatives = FALSE
)
}
\arguments{
\item{InputData}{Name of .xlsx or .csv file that contains input data to be raked.
This file is assumed to have Region (e.g., LHA) by Sex (e.g., 1, 2, 3) as rows, and
Ages (e.g., 0, 1, 2, ..., TOTAL) as columns. Values are population counts.}

\item{Ctrl_Pop_Totals}{Name of .xlsx or .csv file that contains overall control totals
(e.g., "BC AS TOTALS.xlsx"). This file is assumed to have Sex (e.g., 1, 2, 3) as rows and
Ages (e.g., 0, 1, 2, ..., TOTAL) as columns. Values are population counts.
This file typically has dimensions of 3 (obs) by 103 variables.}

\item{Ctrl_Region_Totals}{Name of .xlsx or .csv file that contains overall control totals
(e.g., "LHA TOTALS.xlsx"). Default = NULL. This file is assumed to have Region (e.g., 89 LHAs)
as the first column and TOTAL (population counts) as the second column; this file is not broken out
by Sex or Age. This file typically has dimensions of n (obs) by 2 variables, where "n" is the
number of individual regions (e.g., 89 for LHA). If no name is provided (i.e., NULL), then region
control totals are not used. Instead, the InputData will be used to generate "control" totals.}

\item{Ctrl_AgeGrps_Totals}{Name of .xlsx or .csv file that contains initial 5 year age group totals.
Default = NULL. In virtually all cases, this variable will be NULL. In these cases, the InputData
will be used to generate "control" totals at 5-year groupings (e.g., 0-4, 5-9, 10-14, etc).}

\item{VarRegion}{Name of Region variable in all files (e.g., "LHA")}

\item{VarSex}{Name of Sex variable in all files (e.g., "Sex")}

\item{VarSexTotal}{Value that corresponds to Total (e.g., 3, when 1 and 2 are Male and Female)}

\item{AgeGrpMax}{Age of the older poppulation that will be prorated and raked separately from
other 5 year age groups. AgeGrpMax will include all ages, including itself, through the remainder
of the population. Default = NULL. If AgeGrpMax is not set, the function will use 75 and up
(not necessarily the oldest age; that is, the oldest age is usually 100, meaning 100 and up).
The BC Stats Demographics team determined that 75 was the best age for AgeGrpMax to ensure that
distortion in older populations is minimized.}

\item{allowNegatives}{Logical value for whether or not negative population values are allowed.
Default = FALSE. Only migration data should be allowed to have negative values.}
}
\value{
RakedData.csv will be saved to "outputs" folder, along with various interim files (saved
to "interim_files" within "outputs") and a log file ("raking_log.csv")
}
\description{
Reads a population database file (population, migration) and saves a population database file
with Region values raked for each of Age and Sex. Raking can be run with user-provided region
control totals, or without region control totals. Negative population values may be allowed or
not (default).
This function assumes input files (e.g., InputData, Ctrl_Pop_Totals, etc.) are in an "inputs"
folder, and that the raked output will save to an "outputs" folder. Interim files are also saved
to an "interim_files" folder within "outputs". dbRake() is a large function that takes a few
minutes to run, and depends on multiple smaller functions.
}
\details{
dbRake is a large function with three main parts. Part 1 prorates and rakes Sex values by Region.
Part 2 prorates and rakes 5-year Age Group values by Region and Sex. Part 3 prorates and rakes
single-year Age values by Region and Sex. Throughout, checks are performed and results are
eventually written to raking_log.csv in the "outputs" folder (regardless of whether raking
succeeds or fails). As well, interim files are saved to an "interim_files" folder in "outputs"
for future viewing. If raking succeeds, the final raked data file is saved to "outputs".

dbRake was originally an APL process (not in R). A PDF documenting that process, which holds
true for most of the underlying assumptions and procedures in dbRake, is available on BC Stats'
Y drive (S152\\S52004) in Documentation > Raking > Methodology-Raking_Final.pdf.
}
\seealso{
Raking hekpers include: \code{\link{rounded}}(), \code{\link{read.inputs}}(),
\code{\link{real.to.int}}(), \code{\link{calc.cols}}(), \code{\link{prorate.row}}(),
\code{\link{prep.prorate.col}}(), \code{\link{prorate.col}}(), and raking algorithm functions A, B, C:
\code{\link{allowNegsnoMargin}}(), \code{\link{noNegsnoMargin}}(), \code{\link{noNegsneedMargin}}()
}
\author{
Julie Hawkins, BC Stats
}
