% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{dbRead}
\alias{dbRead}
\title{Read population database}
\usage{
dbRead(db_path, return_format = "default", age_filter = NULL)
}
\arguments{
\item{db_path}{Either vector of database type ("estimates", "projections", "births", "deaths"),
region code and two-character year OR full path to datbase csv file.}

\item{return_format}{Format of returned data.frame. Default is the same columns as default
database format (Year, Type, TypeID, Age, Male, Female, Total), long format gathers sex
as a new column (Year, Type, TypeID, Age, Sex, N), wide format gathers sex as a new column
and spreads ages as columns (Year, Type, TypeID, Sex, 0, 1,... 89, 90+, Total). Allowed
values are "default", "long" or "wide". Default = "default".}

\item{age_filter}{Filters ages based on input vector before returning data in requested format.
Must be a vector. Default = NULL.}
}
\value{
A data.frame object in the requested format.
}
\description{
Reads a population database file (population, births, deaths) in default database format
(Year, Type, TypeID, Age, Male, Female, Total) and returns data. Database path can be either
a full path to a csv database file with the correct columns, or a vector including the type,
region and year of database to be used through \code{\link{getDBPath}}. For example,
db_path = c("estimates", "HA", "18") would read the July 1st 2018 population estimates database
for Local Health Areas. Format of loaded data can be adjusted to default, long or wide and
pre-filtered for specific age groups.For historical reasons, population data is saved as
POPXXEYY (population estimates), POPXXPYY (population projections), BIRXXYY (births) or
DEAXXYY (deaths), where YY is the last two digits of the year. XX represent the shorthand for
the region type. This was passed over from decades of BC Stats PEOPLE development limited to
2-digit codes. Two-digits region codes are as follow:
\itemize{
 \item RD - Regional District, which is the same data as Census Division (CD)
 \item DR - Development region (DR)
 \item HA - Local Heath Area (LHA)
 \item HS - Health Service Delivera Area (HSDA)
 \item HY - Health Authority (HA)
 \item CH - Community Health Service Area (CHSA)
 \item CF - Ministry of Children and Family Development (MCFD)
 \item CA - MCFD Service Delivery Area (MCFD_SDA)
 \item CL - MCFD Local Service Area (MCFD_LSA)
 \item SD - School District (SD)
 \item PS - College Region, or Post-Secondary (CR)
 \item SR - Special Regions
}
}
\seealso{
Other database access helpers: 
\code{\link{dbCheck}()},
\code{\link{dbInfo}()},
\code{\link{dbWrite}()},
\code{\link{getDBPath}()}
}
\author{
Sebastien Lavoie, (formerly, BC Stats)
}
\concept{database access helpers}
