% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.r
\docType{package}
\name{dbutils-package}
\alias{dbutils}
\alias{dbutils-package}
\title{dbutils: Utility functions to aid in database interactions}
\description{
Aids in interacting with databases
}
\details{
\itemize{
\item Connect to databases
\item Capture function calls
\item SQL statement formatting functions
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andybeet/dbutils}
  \item \url{https://andybeet.github.io/dbutils}
  \item Report bugs at \url{https://github.com/andybeet/dbutils/issues}
}

}
\author{
\strong{Maintainer}: Andy Beet \email{andrew.beet@noaa.gov} (\href{https://orcid.org/0000-0001-8270-7090}{ORCID})

}
