% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_db.R
\name{fetch_db}
\alias{fetch_db}
\title{Fetch data from db}
\usage{
fetch_db(conn = NULL, path = NULL, query = NULL, async = FALSE,
  dsn = NULL, ...)
}
\arguments{
\item{conn}{A database connection. Only necessary if `async = FALSE`.}

\item{path}{A SQL file path. Optionally, use `query`.}

\item{query}{A SQL string. Optionally, use `path`. If both are supplied, `path`` is used.}

\item{async}{A logical indicating whether to execute the query asynchronously or (default) not.}

\item{dsn}{The Data Source Name.}

\item{...}{Named values to interpolate into the query string.}
}
\value{
If processed in a synchronous manner (the default), a data frame with as many rows as records and as many
 columns as fields in the result set.
 If processed in an asynchronous manner, a Future. To get the value of a future, use future::value.
 In case of error, a print of the full SQL error is
 returned before execution of error action.
}
\description{
Fetch data from db
}
