% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oracle_tables.R
\name{get_oracle_tables}
\alias{get_oracle_tables}
\title{Extract ALL oracle tables in database}
\usage{
get_oracle_tables(channel, owner = NULL)
}
\arguments{
\item{channel}{DBI Object. Inherited from \link[DBI]{DBIConnection-class}. This object is used to connect
to communicate with the database engine. (see \code{\link[dbutils]{connect_to_database}})}

\item{owner}{character string. The name of the tables owner. Used to filter tables by a specific owner}
}
\value{
Data frame (n x 2)

  \item{column 1}{Nam of the Database}

  \item{column 2}{Name of the table owner}

The default sql statement "\code{"SELECT table_name,owner FROM all_tables ORDER by owner"}" is used
}
\description{
Pulls a list of all of the tables and owners found within a database
}
\section{Reference}{

Use the data dictionary for an html list
}

\examples{
\dontrun{
# extracts ALL table names by owner from the database
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_oracle_tables(channel)


#' # extracts ALL table names by owner = VTR from the database
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_oracle_tables(channel,"VTR")
}

}
\seealso{
\code{\link[dbutils]{connect_to_database}}
}
