% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{getDBPath}
\alias{getDBPath}
\title{Returns a correct full database path}
\usage{
getDBPath(dbType, dbRegion, dbYear)
}
\arguments{
\item{dbType}{Type of database being written. Possible values are "estimates", "projections",
"deaths", "births". Default = "estimates".}

\item{dbRegion}{Two-digit region code of the database. Possible values: "RD", "DR", "HA", "HS",
"HY", "CH", "CF", "CA", "CL", "SD", "PS", "SR" as character. Default = NULL.}

\item{dbYear}{Two-digit year of the data being saved. Based on July 1st reference date,
as character. Default = NULL.}
}
\description{
Returns a population database file (population, births, deaths) path. Database paths are hardcoded
to ensure consistency across population systems. The script uses data type, region type and year
to return an appropriate full path.

The complete list of region IDs and ages for each type of database (population, deaths, births)
can be found in \code{\link{dbutils}}. For historical reasons, population data is saved as
POPRREYY (population estimates), POPRRPYY (population projections), BIRRRYY (births) or
DEARRYY (deaths), where RR is the shorthand for the region code, and YY is the last two digits
of the year.
}
\examples{
getDBPath(dbType = "estimates", dbRegion = "CA", dbYear = "17")  ## "//.../POPCAE17.csv"
getDBPath(dbType = "births", dbRegion = "RD", dbYear = "18")     ## "//.../BIRRD18.csv"
}
\seealso{
Overall package documentation: \code{\link{dbutils}}()

Other database access helpers: 
\code{\link{dbCheck}()},
\code{\link{dbInfo}()},
\code{\link{dbRead}()},
\code{\link{dbWrite}()}
}
\author{
Sebastien Lavoie (formerly, BC Stats)
}
\concept{database access helpers}
