% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRake.R
\name{multiRake}
\alias{multiRake}
\title{Rake over multiple years}
\usage{
multiRake(
  years,
  censusYears = FALSE,
  InputData,
  CtrlPopTotals,
  CtrlRegionTotals = NULL,
  VarRegion,
  change_rake_args = FALSE
)
}
\arguments{
\item{years}{Vector of all years in the data (e.g., Years = 2011:2020).}

\item{censusYears}{Any year(s) that are a census year and should not be raked (e.g., 2011, 2016).
Default is FALSE. If all years need to be raked, set as FALSE.}

\item{InputData}{Name of .xlsx or .csv file that contains input data to be raked.
This file is assumed to have Region (e.g., "TypeID", "CHSA") by Sex (e.g., 1, 2, 3) as rows, and
Ages (e.g., 0, 1, 2, ..., TOTAL) as columns. Values are population counts.}

\item{CtrlPopTotals}{Name of .xlsx or .csv file that contains overall control totals
(e.g., "BC AS TOTALS.xlsx"). This file is assumed to have Sex (e.g., 1, 2, 3) as rows and
Ages (e.g., 0, 1, 2, ..., TOTAL) as columns. Values are population counts.
This file typically has dimensions of 3 (obs) by 103 variables.}

\item{CtrlRegionTotals}{Name of .xlsx or .csv file that contains overall control totals
(e.g., "CHSA TOTALS.xlsx"). Default = NULL. This file is assumed to have Region (e.g., 218 CHSAs)
as the first column and TOTAL (population counts) as the second column; this file is not broken out
by Sex or Age. This file typically has dimensions of n (obs) by 2 variables, where "n" is the
number of individual regions (e.g., 218 for CHSA). If no name is provided (i.e., NULL), then region
control totals are not used. Instead, the InputData will be used to generate "control" totals.}

\item{VarRegion}{Name of Region variable in all files (e.g., "TypeID", "CHSA").}

\item{change_rake_args}{Logical value whether any remaining raking argument defaults need to be
changed. Default = FALSE. If set to TRUE, user will be asked to set the following arguments:
\strong{CtrlAgeGrpsTotals} (default = NULL);
\strong{VarSex} (otherwise pre-specified as "Sex");
\strong{VarSexTotal} (otherwise pre-specified from data's column names);
\strong{AgeGrpMax} (default = NULL which would trigger \code{\link{dbRake}} to use age 75 if
exists; however, multiRake sets this to the strongly recommended age 75);
\strong{allowNegatives} (default = FALSE, should only be TRUE for migration data);
\strong{saveInterimFiles} (default = FALSE);
\strong{writeOutputFile} (default = FALSE);
\strong{writeRakingLog} (default = FALSE);
\strong{readFiles} (default = FALSE which will use files already in environment; if files need
to be read in, set to TRUE).}
}
\value{
RakedData.csv will be saved to "outputs" folder (which will be created if one does not
already exist). If set to TRUE, various interim files will be saved in an "interim_files" folder
within "outputs". If set to TRUE, a log file ("raking_log.csv") will also be saved to the
"outputs" folder.
}
\description{
The raking function, \code{\link{dbRake}}, will run for one year of data but the data often needs
to be raked for more than one year. In \strong{multiRake}, arguments include \strong{Years} in
the data, any \strong{censusYears} that should not be raked as these are generally considered
definitive, name of \strong{InputData}, name of \strong{CtrlPopTotals}, either name of
\strong{CtrlRegionTotals} or default of NULL, name of region variable in all data, and whether
you want to change any arguments required by \code{\link{dbRake}}.
This is a helper function used to run \code{\link{dbRake}} for multiple years.
}
\examples{
\dontrun{  ## if files need to be read in, set 'change_rake_args' to TRUE
           multiRake(years = 2011:2020, censusYears = c(2011, 2016),
                     InputData = "POPHAE19.xlsx", CtrlPopTotals = "BC AS TOTALS.xlsx",
                     CtrlRegionTotals = "LHA TOTALS.xlsx", VarRegion = "LHA",
                     change_rake_args = TRUE) }  ## two census years not to be raked
\dontrun{  multiRake(years = 2012:2016, censusYears = FALSE,
                     InputData = "POPHAE19.xlsx", CtrlPopTotals = "BC AS TOTALS.xlsx",
                     CtrlRegionTotals = "LHA TOTALS.xlsx", VarRegion = "LHA",
                     change_rake_args = FALSE) }   ## all years need to be raked
}
\seealso{
The overall raking function: \code{\link{dbRake}}()

Other raking helpers: 
\code{\link{add.random.fraction.to.cols}()},
\code{\link{adjustSex}()},
\code{\link{allowNegsnoMargin}()},
\code{\link{calc.cols}()},
\code{\link{noNegsneedMargin}()},
\code{\link{noNegsnoMargin}()},
\code{\link{prep.prorate.col}()},
\code{\link{prorate.col}()},
\code{\link{prorate.row}()},
\code{\link{prorate}()},
\code{\link{read.inputs}()},
\code{\link{real.to.int}()},
\code{\link{rename.age.grps}()},
\code{\link{rounded}()}
}
\author{
Julie Hawkins, BC Stats
}
\concept{raking helpers}
