% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRake.R
\name{prorate.col}
\alias{prorate.col}
\title{Adjust data column by column to sum to row totals}
\usage{
prorate.col(CurrCol, n_rowGrps, allowNegatives)
}
\arguments{
\item{CurrCol}{a dataframe of a column to prorate}

\item{n_rowGrps}{the number of groups to adjust over (e.g., number of groups in region)}

\item{allowNegatives}{a logical of whether negative population values are allowed (usually FALSE)}
}
\value{
original dataframe of column, that is now prorated by the adjustment values
}
\description{
Reconcile column by column, when using region control totals, while updating initial estimates of
male/female regional total values if difference is NOT zero, adjust actual data. This function
will be run iteratively through each column of a dataframe already prepared by running through
\code{\link{prep.prorate.col}} while the column's difference is not zero. For example, for 1:n_Sex,
prorate so that region totals sum to region control totals. This is a helper function used in \code{\link{dbRake}}.
}
\seealso{
The overall raking function: \code{\link{dbRake}}()

Other raking helpers: 
\code{\link{add.random.fraction.to.cols}()},
\code{\link{allowNegsnoMargin}()},
\code{\link{calc.cols}()},
\code{\link{multiRake}()},
\code{\link{noNegsneedMargin}()},
\code{\link{noNegsnoMargin}()},
\code{\link{prep.prorate.col}()},
\code{\link{prorate.row}()},
\code{\link{read.inputs}()},
\code{\link{real.to.int}()},
\code{\link{rename.age.grps}()},
\code{\link{rounded}()}
}
\concept{raking helpers}
