% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRake.R
\name{prorate}
\alias{prorate}
\title{Adjust data to sum to control totals via prorate.row}
\usage{
prorate(data, control, VarRow, allowNegatives = FALSE)
}
\arguments{
\item{data}{a dataframe to prorate (one row per VarRow)}

\item{control}{a dataframe of control totals (one row per VarRow)}

\item{VarRow}{a character, the name of the variable to prorate on (e.g., when prorating Sex for LHA to region control totals, VarRow = "Region")}

\item{allowNegatives}{a logical of whether negative population values are allowed (usually FALSE)}
}
\value{
original dataframe, that is now prorated by the adjustment values
}
\description{
Prorating is for adjusting data to control totals in only 1 dimension. Use dbRake to adjust data in 2 dimensions.
}
\details{
This function iteratively adjusts each row of data by running \code{\link{prorate.row}} which
first prepares rows by running through \code{\link{calc.cols}} while the row's difference is not zero.
For example, for 1:n_Sex, prorate
so that region totals sum to region control totals. This is a helper function used in \code{\link{dbRake}}.
}
\note{
This function not currently integrated into dbRake
}
\seealso{
The overall raking function: \code{\link{dbRake}}()

Other raking helpers: 
\code{\link{add.random.fraction.to.cols}()},
\code{\link{adjustSex}()},
\code{\link{allowNegsnoMargin}()},
\code{\link{calc.cols}()},
\code{\link{multiRake}()},
\code{\link{noNegsneedMargin}()},
\code{\link{noNegsnoMargin}()},
\code{\link{prep.prorate.col}()},
\code{\link{prorate.col}()},
\code{\link{prorate.row}()},
\code{\link{read.inputs}()},
\code{\link{real.to.int}()},
\code{\link{rename.age.grps}()},
\code{\link{rounded}()}
}
\concept{raking helpers}
