% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRake.R
\name{prorate.row}
\alias{prorate.row}
\title{Adjust data row by row to sum to control totals}
\usage{
prorate.row(CurrRow, n_colGrps, allowNegatives)
}
\arguments{
\item{CurrRow}{a dataframe of a row to prorate}

\item{n_colGrps}{the number of groups to adjust over (e.g., number of groups in region)}

\item{allowNegatives}{a logical of whether negative population values are allowed (usually FALSE)}
}
\value{
original dataframe of row, that is now prorated by the adjustment values
}
\description{
Reconcile row by row, when using region control totals, while updating initial estimates of
male/female regional total values if difference is NOT zero, adjust actual data. This function
will be run iteratively through each row of a dataframe already prepared by running through
\code{\link{calc.cols}} while the row's difference is not zero. For example, for 1:n_Sex, prorate
so that region totals sum to region control totals. This is a helper function used in \code{\link{dbRake}}.
}
\seealso{
The overall raking function: \code{\link{dbRake}}()

Other raking helpers: 
\code{\link{add.random.fraction.to.cols}()},
\code{\link{adjustSex}()},
\code{\link{allowNegsnoMargin}()},
\code{\link{calc.cols}()},
\code{\link{multiRake}()},
\code{\link{noNegsneedMargin}()},
\code{\link{noNegsnoMargin}()},
\code{\link{prep.prorate.col}()},
\code{\link{prorate.col}()},
\code{\link{prorate}()},
\code{\link{read.inputs}()},
\code{\link{real.to.int}()},
\code{\link{rename.age.grps}()},
\code{\link{rounded}()}
}
\concept{raking helpers}
