% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRake.R
\name{real.to.int}
\alias{real.to.int}
\title{Converts real numbers to integers}
\usage{
real.to.int(realNums)
}
\arguments{
\item{realNums}{a  vector of real (fraction) number}
}
\value{
a vector of integer numbers that sums to the rounded sum of realNums
}
\description{
Converts real numbers (fractions) to integer numbers while preserving rounded sum.
This is a helper function used in \code{\link{dbRake}}.
}
\seealso{
The overall raking function: \code{\link{dbRake}}()

Other raking helpers: 
\code{\link{add.random.fraction.to.cols}()},
\code{\link{adjustSex}()},
\code{\link{allowNegsnoMargin}()},
\code{\link{calc.cols}()},
\code{\link{multiRake}()},
\code{\link{noNegsneedMargin}()},
\code{\link{noNegsnoMargin}()},
\code{\link{prep.prorate.col}()},
\code{\link{prorate.col}()},
\code{\link{prorate.row}()},
\code{\link{read.inputs}()},
\code{\link{rename.age.grps}()},
\code{\link{rounded}()}
}
\concept{raking helpers}
