% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRake.R
\name{rename.age.grps}
\alias{rename.age.grps}
\title{Rename negative columns as 5 year age groups}
\usage{
rename.age.grps(data, VarRegion, VarSex)
}
\arguments{
\item{data}{A dataset with -X1, -X2, -X3, ... column names}

\item{VarRegion}{Name of Region variable in all files (e.g., "LHA")}

\item{VarSex}{Name of Sex variable in all files (e.g., "Sex")}
}
\value{
data with X-Y col names
}
\description{
Some databases may have 5 year age groups as -4, -9, -14, ... meaning 0-4, 5-9, 10-14, ....
However, the \code{\link{dbRake}} function requires depends on age group column names as 'X-Y'.
This is a helper function used in \code{\link{dbRake}}.
}
\seealso{
The overall raking function: \code{\link{dbRake}}()

Other raking helpers: 
\code{\link{add.random.fraction.to.cols}()},
\code{\link{allowNegsnoMargin}()},
\code{\link{calc.cols}()},
\code{\link{multiRake}()},
\code{\link{noNegsneedMargin}()},
\code{\link{noNegsnoMargin}()},
\code{\link{prep.prorate.col}()},
\code{\link{prorate.col}()},
\code{\link{prorate.row}()},
\code{\link{read.inputs}()},
\code{\link{real.to.int}()},
\code{\link{rounded}()}
}
\concept{raking helpers}
