% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRake.R
\name{rounded}
\alias{rounded}
\title{Round a number}
\usage{
rounded(x)
}
\arguments{
\item{x}{Number to be rounded}
}
\value{
Rounded number
}
\description{
For some reason, R rounds to even (i.e., 12.5 rounds to 12), but we want to round to 0
(i.e., 12.5 rounds to 13). This is a helper function used in \code{\link{dbRake}}.
}
\examples{
rounded(1.2345)  ## 1
rounded(12.5)    ## 13
}
\seealso{
The overall raking function: \code{\link{dbRake}}()

Other raking helpers: 
\code{\link{add.random.fraction.to.cols}()},
\code{\link{allowNegsnoMargin}()},
\code{\link{calc.cols}()},
\code{\link{noNegsneedMargin}()},
\code{\link{noNegsnoMargin}()},
\code{\link{prep.prorate.col}()},
\code{\link{prorate.col}()},
\code{\link{prorate.row}()},
\code{\link{read.inputs}()},
\code{\link{real.to.int}()},
\code{\link{rename.age.grps}()}
}
\author{
Stephanie Yurchak, BC Stats
}
\concept{raking helpers}
