% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_sql_query_nespp3.R
\name{split_sql_query_nespp3}
\alias{split_sql_query_nespp3}
\title{Splits a query into multiple parts}
\usage{
split_sql_query_nespp3(channel, NESPP4s, maxItems = 900)
}
\arguments{
\item{channel}{an Object inherited from \link[DBI]{DBIConnection-class}. This object is used to connect
to communicate with the database engine. (see \code{\link{connect_to_database}})}

\item{NESPP4s}{Numerical Vector. NESPP4 codes used in an sql query}

\item{maxItems}{Numeric scalar. Maximum number of items in a single sql query}
}
\value{
List
\item{sql}{the sql statements}
\item{speciesTable}{Resulting table from concatenating the sql calls}
}
\description{
If a query has > 1000 items it will fail. For all cases where a query has > \code{maxItems} the call will be split. This occurrs when NESPP4 codes used in an a sql call.
}
