% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.R
\name{sql_array}
\alias{sql_array}
\title{Transform a vector into a SQL-interpreted array.}
\usage{
sql_array(vector)
}
\arguments{
\item{vector}{character vector}
}
\value{
List with a single string that's interpreted as a SQL array to interpolate in a query.
}
