% Generated by roxygen2 (4.1.1): do not edit by hand
\docType{data}
\name{sql_NULL}
\alias{sql_NOW}
\alias{sql_NULL}
\title{SQL Special Types}
\format{\preformatted{  expression("NULL")
}}
\usage{
sql_NULL

sql_NOW
}
\description{
  These variables can be used in SQL statements defined by \code{\link{SQLquery}} objects.
}
\keyword{datasets}

