% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConvert.R
\name{conversionRead}
\alias{conversionRead}
\title{Read population conversion table}
\usage{
conversionRead(conv_path)
}
\arguments{
\item{conv_path}{blah blah}
}
\value{
blah blah
}
\description{
lorem ipsum
}
\seealso{
Other conversion helpers: 
\code{\link{dbConvert}()}
}
\concept{conversion helpers}
