% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_species_lookup.R
\name{create_species_lookup}
\alias{create_species_lookup}
\title{Pair NESPP3, SVSPP3, Species_itis etc}
\usage{
create_species_lookup(channel, species = "all", speciesType = "NESPP3")
}
\arguments{
\item{channel}{an Object inherited from \link[DBI]{DBIConnection-class}. This object is used to connect
to communicate with the database engine. (see \code{\link{connect_to_database}})}

\item{species}{Character String. Either NESPP3, SVSPP, Species_itis (not case sensitive)}

\item{speciesType}{Character string. Type of species code entered for \code{species}. Default = "NESPP3"}
}
\value{
Tibble:

The number of rows = the length of \code{species} with columns
COMMON_NAME, SCIENTIFIC_NAME, SPECIES_ITIS, NESPP3, NAFOSPP, SVSPP
}
\description{
Create a table with all species infomation across databases. 
Combine  NESPP3, SVSPP3, species_itis, common name and scientific name
}
\section{Database tables used}{


Two tables are used to compile infomation. cfdbs.cfspp and cfdbs.species_itis_ne.
The resulting tables are then joined.
}

