% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_table.R
\name{db_table}
\alias{db_table}
\title{Create a tbl that references a schema.table in the `conn` database.}
\usage{
db_table(conn, schema, table)
}
\arguments{
\item{conn}{A database connection.}

\item{schema}{A string with schema name.}

\item{table}{A string with table name.}
}
\value{
A reference to a database table.
}
\description{
Create a tbl that references a schema.table in the `conn` database.
}
