% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{dbCheck}
\alias{dbCheck}
\title{Check population database for errors}
\usage{
dbCheck(db, full_BC = TRUE)
}
\arguments{
\item{db}{Data variable containing the database to be checked.}

\item{full_BC}{WHAT THE HECK IS THIS? Default = TRUE.}
}
\value{
db_ok, a logical. If the database passes all checks, db_ok = TRUE. Otherwise, any error(s)
will be printed to screen and db_ok will be set to FALSE.
}
\description{
Checks a database for specific errors. (1) Database must have 7 columns: Year, Type, TypeID, Age,
Male, Female, and Total. (2) Each Year of data must have the same number of occurrences.
(3) Years of data must be continuous (no year can be missing). (4) The Total column must equal
the sum of the Male and Female columns. (5) THIS FUNCTION IS INCOMPLETE.
}
\seealso{
Other database access helpers: 
\code{\link{dbInfo}()},
\code{\link{dbRead}()},
\code{\link{dbWrite}()},
\code{\link{getDBPath}()}
}
\author{
Sebastien Lavoie, (formerly, BC Stats); Julie Hawkins, BC Stats (remaining)
}
\concept{database access helpers}
