% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConvert.R
\name{dbConvert}
\alias{dbConvert}
\title{Convert population database}
\usage{
dbConvert(
  db,
  conv_table,
  years = NULL,
  rake = TRUE,
  change_rake_args = FALSE,
  full_BC = TRUE
)
}
\arguments{
\item{db}{Data variable containing the source database to be converted. Expects data to be in
data.frame with columns: Year, Type, TypeID, Age, Male, Female, Total.}

\item{conv_table}{Name of conversion table to be read in to provide the proportional
relationship ("split") between source and destination geographies. Conversion table path is
hardcoded to ensure consistency across population systems, and will find conversion table in
".../ConversionTables/" folder (on the LAN).}

\item{years}{Vector of one or more years to be included in the converted destination data.
Default = NULL. If NULL, the function will convert for all years in the source data. If not NULL,
the function will convert only the year(s) specified, but return all years in source data (which
can be confusing; therefore, it is recommended to include in db only the year(s) to be converted.)}

\item{rake}{Whether raking is required (default) or not. Default = TRUE.
While the \code{\link{dbRake}} function requires InputData and CtrlPopTotals (either already
in the environment or to be read in from an "inputs" folder), `dbConvert` will create these
from "db" and the conversion table, so does not need these arguments specified.}

\item{change_rake_args}{Logical value whether raking argument defaults need to be changed.
Default = FALSE. If set to TRUE, user will be asked to set the following arguments:
\itemize{
  \item \strong{CtrlAgeGrpsTotals} (default = NULL);
  \item \strong{VarRegion} (otherwise pre-specified as From geography found in name of the conversion
table, e.g., HA);
  \item \strong{VarSex} (otherwise pre-specified as "Sex");
  \item \strong{VarSexTotal} (otherwise pre-specified from data's column names);
  \item \strong{AgeGrpMax} (default = NULL which would trigger \code{\link{dbRake}} to use age 75 if
exists; however, dbConvert sets this to the strongly recommended age 75);
  \item \strong{allowNegatives} (default = FALSE, should only be TRUE for migration data);
  \item \strong{saveInterimFiles} (default = FALSE);
  \item \strong{writeOutputFile} (default = FALSE);
  \item \strong{writeRakingLog} (default = FALSE).
}
Regardless of `change_rake_args` value, \code{\link{dbRake}}'s argument `readFiles` will be
set to FALSE because the \strong{input files will be created during the conversion process}.}

\item{full_BC}{Logical value whether the region covers all of BC. Default = TRUE. Those regions
(e.g., VI, CMAs) that do not cover all of BC should have full_BC = FALSE, and their sum is not
checked against the BC total. As well, during conversion, source geographies not included in the
destination geographies are dropped from the From and To working datafiles.}
}
\value{
Database converted from source to destination geography. If not all allocations were 100,
(i.e., some splits < 100), then raking was also done (unless explicitly set to FALSE). Note that
Age can be a positive age (e.g., 0, 1, 2, ...), a 5-year age group (e.g., 0-4) or TOTAL.
}
\description{
Converts a population database to another geography using a conversion table and, possibly,
raking.
}
\details{
Conversion tables should be in the ".../ConversionTables/" LAN folder so that they can be found,
and are expected to be .xlsx files with three unnamed columns with the proportional relationship
(aka, "split"), source and destination geographies.

The conversion process is as follows:
\enumerate{
 \item Evaluate the conversion table between Geography A (source)and B (destination) for regions
 that require no conversion (i.e., regions are converted 1-to-1 from Geography A to B). Remove
 these regions from the conversion process.
 \item Sequentially share out the data relating to the region in Geography A using the
 conversion factor to the respective region(s) in Geography B.
 \item Aggregate Geography B components that received contributions from multiple regions in
 Geography A.
 \item If not all regions are 1-to-1, "split" destination geographies must conform to source
 totals (aka, "control population totals"), achieved using raking (\code{\link{dbRake}}).
 If Step (1) removed some 1-to-1 regions, the control totals will be adjusted downward. Destination
 control totals (aka, "control region totals", or dbRake's "CtrlRegionTotals") are set as NULL and
 will be generated during the raking process from the converted destination data (aka, "InputData").
 \item Any 1-to-1 regions removed during Step (1) are added back to Geography B (under the
 Geography B naming convention).
}
}
\examples{
\dontrun{ dbConvert(db = "POPHAP20", conv_table = "Table-LHA-HSDA-2019.xlsx", rake = FALSE) } ## all 100\%
\dontrun{ dbConvert(db = "POPHAP20", conv_table = "Table-LHA-SD-2019") }          ## some split
\dontrun{ dbConvert(db = "POPCSE21", conv_table = "Table-CSD-VI_NO_CRD-2020", full_BC = FALSE) }
}
\seealso{
Overall package documentation: \code{\link{dbutils}}()

Other conversion helpers: 
\code{\link{conversionRead}()},
\code{\link{conversionTables}()},
\code{\link{geogConvert}()}
}
\concept{conversion helpers}
