% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConvert.R
\name{dbConvert}
\alias{dbConvert}
\title{Convert population database}
\usage{
dbConvert(db, conv_table, rake = FALSE, region_total, control_total)
}
\arguments{
\item{db}{blah blah}

\item{conv_table}{lorem ipsum}

\item{rake}{blah Default = FALSE.}

\item{region_total}{lorem ipsum dolor}

\item{control_total}{sit amen}
}
\value{
dolor sit amen
}
\description{
lorem ipsum
}
\seealso{
Other conversion helpers: 
\code{\link{conversionRead}()}
}
\concept{conversion helpers}
