% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{dbRead}
\alias{dbRead}
\title{Read population database}
\usage{
dbRead(db_path, return_format = "default", age_filter = NULL)
}
\arguments{
\item{db_path}{Either vector of database type ("estimates", "projections", "births", "deaths"),
region code and two-character year OR full path to database csv file.}

\item{return_format}{Format of returned data.frame. Default is the same columns as default
database format (Year, Type, TypeID, Age, Male, Female, Total), long format gathers sex
as a new column (Year, Type, TypeID, Age, Sex, N), wide format gathers sex as a new column
and spreads ages as columns (Year, Type, TypeID, Sex, 0, 1,... 89, 90+, Total). Allowed
values are "default", "long" or "wide". Default = "default".}

\item{age_filter}{Filters ages based on input vector before returning data in requested format.
Must be a vector. Default = NULL.}
}
\value{
A data.frame object in the requested format.
}
\description{
Reads a population database file (population, births, deaths) in default database format
(Year, Type, TypeID, Age, Male, Female, Total) and returns data. Database path can be either
a full path to a csv database file with the correct columns, or a vector including the type,
region and year of database to be used through \code{\link{getDBPath}}. For example,
db_path = c("estimates", "HA", "19") would read the July 1st 2019 population estimates database
for Local Health Areas. Format of loaded data can be adjusted to default, long or wide and
pre-filtered for specific age groups.

Note that \code{\link{dbCheck}} is run before reading the file. User will be warned if any database
checks fail.
}
\details{
The complete list of region IDs and ages for each type of database (population, deaths, births)
can be found in \code{\link{dbutils}}. For historical reasons, population data is saved as
POPRREYY (population estimates), POPRRPYY (population projections), BIRRRYY (births) or
DEARRYY (deaths), where RR is the shorthand for the region code, and YY is the last two digits
of the year.
}
\examples{
dbRead(c("estimates", "HA", "19"), return_format = "wide", age_filter = c(0, 1, 2, 3, 4, -999))
\dontrun{   dbRead("I://PopulationR/Database/Estimates/POPHAE19.csv", return_format = "wide",
                   age_filter = c(0, 1, 2, 3, 4, -999))  }
}
\seealso{
Overall package documentation: \code{\link{dbutils}}()

Other database access helpers: 
\code{\link{dbCheck}()},
\code{\link{dbInfo}()},
\code{\link{dbWrite}()},
\code{\link{getDBPath}()}
}
\author{
Sebastien Lavoie (formerly, BC Stats)
}
\concept{database access helpers}
