% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.R
\docType{package}
\name{dbutils}
\alias{dbutils}
\title{dbutils: A package of utility functions for processing population data}
\description{
The dbutils package provides three categories of important functions:
database access functions (getDBPath, dbCheck, dbInfo, dbRead, dbWrite),
conversion functions (conversionRead, dbConvert),
and raking functions (dbRake, raking helper functions).
}
\section{dbutils functions}{

The dbutils functions are used to process population data.
}

