% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{dbWrite}
\alias{dbWrite}
\title{Write population database}
\usage{
dbWrite(db, db_path, overwrite = FALSE)
}
\arguments{
\item{db}{Data variable containing the database to be written. Expects data to be in data.frame
with columns: Year, Type, TypeID, Age, Male, Female, Total.}

\item{db_path}{Either vector of database type ("estimates", "projections", "births", "deaths"),
region code and two-character year OR full path to datbase csv file.}

\item{overwrite}{Whether to overwrite the file if a database already exists. Default = FALSE.}
}
\description{
Writes a population database file (population, births, deaths). Data format is kept consistent
with columns always including Year, Type, TypeID, Age, Male, Female, Total, in this order. As
well, population databases include all possible region IDs and ages, even if that means a
Male/Female/Total with N = 0. The complete list of region IDs and ages for each type of
database (population, deaths, births) can be found at the \code{\link{dbutils}}. For historical
reasons, population data is saved as POPXXEYY (population estimates), POPXXPYY (population projections),
BIRXXYY (births) or DEAXXYY (deaths), where YY is the last two digits of the year. XX represents
the shorthand for the region type. This was passed over from decades of BC Stats PEOPLE development
limited to 2-digit codes.
Two-digits region codes are as follow:
\itemize{
 \item RD - Regional District, which is the same data as Census Division (CD)
 \item DR - Development region (DR)
 \item HA - Local Heath Area (LHA)
 \item HS - Health Service Delivera Area (HSDA)
 \item HY - Health Authority (HA)
 \item CH - Community Health Service Area (CHSA)
 \item CF - Ministry of Children and Family Development (MCFD)
 \item CA - MCFD Service Delivery Area (MCFD_SDA)
 \item CL - MCFD Local Service Area (MCFD_LSA)
 \item SD - School District (SD)
 \item PS - College Region, or Post-Secondary (CR)
 \item SR - Special Regions
}
Database output path does not have to be specified. The paths are hard-coded to ensure data is
kept in the correct location. All that is needed is a database type (population estimates,
population projections, deaths or births), the two-digit region code and the last two-digits of
the database year based on a July 1st reference date. This is enough to figure out the full path
of the database through the \code{\link{getDBPath}} function.
}
\seealso{
Other database access helpers: 
\code{\link{dbCheck}()},
\code{\link{dbInfo}()},
\code{\link{dbRead}()},
\code{\link{getDBPath}()}
}
\concept{database access helpers}
