% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.R
\name{fetch_db}
\alias{fetch_db}
\title{Fetch data from db}
\usage{
fetch_db(conn, path, ...)
}
\arguments{
\item{conn}{A database connection.}

\item{path}{A sql file path.}

\item{...}{Named values to interpolate into the query string.}
}
\value{
A data frame with as many rows as records and as many
 columns as fields in the result set.
}
