% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAccess.R
\name{getDBPath}
\alias{getDBPath}
\title{Returns a correct full database path}
\usage{
getDBPath(dbType, dbRegion, dbYear)
}
\arguments{
\item{dbType}{Type of database being written. Possible values are "estimates", "projections",
"deaths", "births". Default = "estimates".}

\item{dbRegion}{Two-digit region code of the database. Possible values: "RD", "DR", "HA", "HS",
"HY", "CH", "CF", "CA", "CL", "SD", "PS", "SR" as character. Default = NULL.}

\item{dbYear}{Two-digit year of the data being saved. Based on July 1st reference date,
as character. Default = NULL.}
}
\description{
Returns a population database file (population, births, deaths) path. Database paths are hardcoded
to ensure consistency across population systems. The script uses data type, region type and year
to return an appropriate full path. The complete list of region IDs and ages for each type of
database (population, deaths, births) can be found in the dbutils.R script. For historical reasons,
population data is saved as POPXXEYY (population estimates), POPXXPYY (population projections),
BIRXXYY (births) or DEAXXYY (deaths), where YY is the last two digits of the year. XX represent
the shorthand for the region type. This was passed over from decades of BC Stats PEOPLE
development limited to 2-digit codes. Two-digits region codes are as follow:
\itemize{
 \item RD - Regional District, which is the same data as Census Division (CD)
 \item DR - Development region (DR)
 \item HA - Local Heath Area (LHA)
 \item HS - Health Service Delivera Area (HSDA)
 \item HY - Health Authority (HA)
 \item CH - Community Health Service Area (CHSA)
 \item CF - Ministry of Children and Family Development (MCFD)
 \item CA - MCFD Service Delivery Area (MCFD_SDA)
 \item CL - MCFD Local Service Area (MCFD_LSA)
 \item SD - School District (SD)
 \item PS - College Region, or Post-Secondary (CR)
 \item SR - Special Regions
}
}
\examples{
getDBPath(dbType = "estimates", dbRegion = "CA", dbYear = "17") ## "//SFP.IDIR.BCGOV/S152/S52007/PopulationR/Database/Estimates/POPCAE17.csv"
getDBPath(dbType = "births", dbRegion = "RD", dbYear = "18") ## "//SFP.IDIR.BCGOV/S152/S52007/VITAL/Database/Births/BIRRD18.csv"
}
\seealso{
Other database access helpers: 
\code{\link{dbCheck}()},
\code{\link{dbInfo}()},
\code{\link{dbRead}()},
\code{\link{dbWrite}()}
}
\concept{database access helpers}
